requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

crewai==0.11.2
langchain-google-genai==1.0.1
openai==1.16.2
pygame==2.5.2
python-dotenv==1.0.0
streamlit==1.33.0
